requirements/requirements.txt
# This file is expected by Heroku.
SQLAlchemy~=2.0.35
python-dotenv
pandas~=2.2.3
django-debug-toolbar
django-extensions
djangorestframework
gunicorn>=23.0.0                       #+ https://github.com/benoitc/gunicorn
#psycopg2==2.8.6 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
Collectfast==2.2.0                     # https://github.com/antonagestam/collectfast
django-anymail[mailgun]>=12.0          #+
django-redis~=5.4.0                    #+ https://github.com/jazzband/django-redis

Werkzeug>=3.1.3                        #+ https://github.com/pallets/werkzeug
ipdb>=0.13.13                          #+ https://github.com/gotcha/ipdb
psycopg2-binary>=2.9.10                #+https://github.com/psycopg/psycopg2
openpyxl

# Testing
# ------------------------------------------------------------------------------
mypy>=1.14.1 # https://github.com/python/mypy
# mypy >0.770 and <0.780 is required by django-stubs==1.5.0
django-stubs>=5.1.1                    #+ https://github.com/typeddjango/django-stubs
pytest>=8.3.4                          #+ https://github.com/pytest-dev/pytest
pytest-sugar>=1.0.0                    #+ https://github.com/Frozenball/pytest-sugar
django-test-plus>=2.2.4                #+ https://github.com/revsys/django-test-plus

# Code quality
# ------------------------------------------------------------------------------
flake8>=7.1.1                          #+ https://github.com/PyCQA/flake8
coverage>=7.6.10                       #+ https://github.com/nedbat/coveragepy
black==24.10.0                         #+ https://github.com/ambv/black
pylint-django>=2.6.1                   #+ https://github.com/PyCQA/pylint-django
pre-commit>=4.0.1                      #+ https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy>=3.3.1                     #+ https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.1              #+ https://github.com/jazzband/django-debug-toolbar
django-extensions~=3.2.3               #+ https://github.com/django-extensions/django-extensions
django-coverage-plugin>=3.1.0          #+ https://github.com/nedbat/django_coverage_plugin
pytest-django>=4.9.0                   #+ https://github.com/pytest-dev/pytest-django

pytz>=2024.2                           #+ https://github.com/stub42/pytz
python-slugify>=8.0.4                  #+ https://github.com/un33k/python-slugify
Pillow>=11.1.0                         #+ https://github.com/python-pillow/Pillow
argon2-cffi                            # https://github.com/hynek/argon2_cffi
whitenoise>=6.8.2                      #+ https://github.com/evansd/whitenoise
unidecode>=1.3.8                       #+ https://pypi.org/project/Unidecode/
setuptools>=75.7.0                     #+

# Django
# ------------------------------------------------------------------------------
django~=4.2 # https://www.djangoproject.com/
django-environ>=0.11.2                 #+ https://github.com/joke2k/django-environ
django-model-utils                     #+ https://github.com/jazzband/django-model-utils
django-allauth>=65.3.1                 #+ https://github.com/pennersr/django-allauth
django-crispy-forms                    #+ https://github.com/django-crispy-forms/django-crispy-forms
django-autoslug>=1.9.9                 #+ https://github.com/justinmayer/django-autoslug/
django-heroku



